Refactor code into channel_output func
